Fixed issues with allocators
I had a been using pointers into the linear allocator which would
break every time the allocator would do a reallocation to adjust
it's size.
The new solution is to use Slices. These are just a position into
the allocator and a size. So when the allocator does a
reallocation, the Slices still point to the same piece of information.